move storage adapter
