+++ package.json
@@ -932,6 +932,7 @@
     "@kbn/status-plugin-a-plugin": "link:test/server_integration/plugins/status_plugin_a",
     "@kbn/status-plugin-b-plugin": "link:test/server_integration/plugins/status_plugin_b",
     "@kbn/std": "link:src/platform/packages/shared/kbn-std",
+    "@kbn/storage-adapter": "link:src/platform/packages/shared/kbn-storage-adapter",
     "@kbn/streams-app-plugin": "link:x-pack/solutions/observability/plugins/streams_app",
     "@kbn/streams-plugin": "link:x-pack/solutions/observability/plugins/streams",
     "@kbn/streams-schema": "link:x-pack/solutions/observability/packages/kbn-streams-schema",
