+++ src/platform/packages/shared/kbn-storage-adapter/README.md
