+++ tsconfig.base.json
@@ -1894,6 +1894,8 @@
       "@kbn/std/*": ["src/platform/packages/shared/kbn-std/*"],
       "@kbn/stdio-dev-helpers": ["packages/kbn-stdio-dev-helpers"],
       "@kbn/stdio-dev-helpers/*": ["packages/kbn-stdio-dev-helpers/*"],
+      "@kbn/storage-adapter": ["src/platform/packages/shared/kbn-storage-adapter"],
+      "@kbn/storage-adapter/*": ["src/platform/packages/shared/kbn-storage-adapter/*"],
       "@kbn/storybook": ["packages/kbn-storybook"],
       "@kbn/storybook/*": ["packages/kbn-storybook/*"],
       "@kbn/streams-app-plugin": ["x-pack/solutions/observability/plugins/streams_app"],
