--- x-pack/solutions/observability/packages/utils-server/es/storage/get_schema_version.ts
@@ -1,15 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-import stringify from 'json-stable-stringify';
-import objectHash from 'object-hash';
-import { IndexStorageSettings } from '.';
-
-export function getSchemaVersion(storage: IndexStorageSettings): string {
-  const version = objectHash(stringify(storage.schema.properties));
-  return version;
-}
